libass: fix type mismatch between size parameter and the way it's used
git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@27531 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ libass/ass.c
@@ -1021,7 +1021,7 @@ ass_track_t* ass_read_memory(ass_library_t* library, char* buf, size_t bufsize, 
 	return track;
 }
 
-char* read_file_recode(char* fname, char* codepage, int* size)
+char* read_file_recode(char* fname, char* codepage, size_t* size)
 {
 	char* buf;
 	size_t bufsize;
